+++ doc/changelog.rst
@@ -6,6 +6,7 @@ Version 0.7.9
 
 * metadata: add ``nohz_full`` and ``isolated`` to ``cpu_config``
 * add ``--affinity`` option to the ``metadata`` command
+* Add :meth:`Benchmark.get_unit` method
 
 Version 0.7.8 (2016-09-10)
 --------------------------